TP 7 - OOP
- enum class : 
Is a special type of class used to define a fixed set of constants. 
Each constant in an enum class is an object.

- data class : 
Is a special type of class that is designed to hold data.

- sealed class : 
Is a special kind of class that restricts which other classes can inherit from it.